Initialise basic Prometheus+Grafana stack
I want to be able to a) collect data and b) graph it. This configures
Grafana (for b) with a Prometheus datasource (for a) using Docker
Compose.

Compose, while I wouldn't use it in a "proper" production environment,
is nice and lightweight (in terms of use), meaning I can get started
quickly and have everything documented in code.

For the moment, I'm not thinking too hard about if I want both of these
exposed on the network or not, or how I want them exposed (bridging etc)
because I mostly just want to get stuff working.