Monitor local and upstream DNS performance
We're using the Prometheus-built [Blackbox probe
exporter](https://github.com/prometheus/blackbox_exporter) to monitor
our ability to use DNS (and soon HTTP).

We want to monitor responses both from the local DNS server on our
network, and directly from our upstream DNS resolvers (which, right
now, are Cloudflare's 1.1.1.1 and 1.0.0.1) to ensure that DNS traffic
is flowing both within and out of the network correctly.

I followed [this guide](https://www.robustperception.io/probing-dns-servers-with-the-blackbox-exporter/),
but somewhat simplified the configuration because I'm happy with most
of the defaults for things they specify. For example, the DNS probe
defaults to requesting ANY record, and we want to just ensure we get
results for something that's relatively unlikely to have its DNS
entirely disappear (hence: Google).