Monitor HTTP traffic availability
We've been having occasional dropouts with our network recently, and
we'd like to understand if these are wifi, the network, or the ISP.

This change uses Blackbox to monitor a few of our more heavily-used
services' primary domains just to make sure we can still theoretically
reach them. It's mostly based on a mix of the config from [this blog
post](https://mostly-bsd.github.io/post/monitoring/) and [Blackbox's
own sample config](https://github.com/prometheus/blackbox_exporter/blob/master/example.yml).

This change also tweaks the job names because we don't actually care at
the point of use that Blackbox is what's doing the monitoring, just that
it's the DNS/HTTP monitor.