Add a basic network dashboard
We want easy visibility of the speed of our internet connection and its
quality. This should give us an idea of the upload and download speed
trends over time, as well as the quality of both DNS and HTTP
connections. That way, we should be able to get some idea of where
problems might lie, and where to look in terms of debugging.