Run on Docker-incompatible hosts using Vagrant
My target environment for running this (for now) is an old iMac running
macOS 10.13.6 (High Sierra). Docker Desktop for Mac requires macOS 11
(Big Sur; 3 releases later) or newer, so we can't run it.

This works around that by running Docker inside a VM (which is
ultimately what Docker Desktop does anyway) using Vagrant. This also
allows us to bridge the VM to the external network (which Docker Desktop
doesn't support, per https://github.com/docker/for-mac/issues/3926).

There are a few provisos to this:
- It assumes you're using Virtualbox
- The network bridge is currently hard-coded with the assumption that
  you're bridging to a device called `en0: Ethernet`. Which *I* am, but
  others may not be.
- None of this has been tested outside my target machine, so may break.