Run speed tests elsewhere for now
After much tinkering, it turns out that I just can't get in-VM
speedtests to be any faster than about 2/3 of the speed on the host.

Given that, I've decided to run the speed tests on the host, given
spinning up an instance of https://github.com/MiguelNdeCarvalho/speedtest-exporter
isn't that hard. I've left all the stuff configured in this repo,
though, because I will probably want to continue to tinker with it over
time, and it does no harm being there. (Famous last words!)