Make speedtest gauges combine all results
We always want to present a unified view of all download/upload results,
regardless of their source (in the headline gauges, anyway). This
extracts the label properties to fields, merges the results from all
targets (so we end up with one download and one upload result), then
sorts them by time (so the “last” item is indeed the last).

It also extends the default window of the dashboard to the last 12 hours.