Ensure containers start with the VM
We always want the containers to start when the VM does. Docker allows
us to use a restart policy of `always` to [start the containers whenever
the Docker daemon starts](https://docs.docker.com/config/containers/start-containers-automatically/).

That way, we can stop/update them when we need to, but the Docker
daemon will always restart them when it's running (for example, when it
starts after a reboot of the VM).

This also removes an unnecessary `apt-get update` from the VM
provisioning process.